Move footer, globals, navbar, pagination components to app
