+++ app/account/_components/ResetPasswordForm.tsx
@@ -4,7 +4,7 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
 
 import Button from 'app/_components/button/Button';
-import { Spinner } from '~/components/globals/Spinner';
+import { Spinner } from 'app/_components/globals/Spinner';
 
 import Container from '~/layouts/Container';
 import swell from '~/lib/SwellJS';
