+++ app/account/addresses/_components/DeleteAddressModal.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { GrClose } from 'react-icons/gr';
 
 import Button from 'app/_components/button/Button';
+import { Spinner } from 'app/_components/globals/Spinner';
 import Modal from 'app/account/_components/Modal';
-import { Spinner } from '~/components/globals/Spinner';
 
 import swell from '~/lib/SwellJS';
