+++ app/account/reset-password/ResetPasswordForm.tsx
@@ -6,7 +6,7 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
 
 import Button from 'app/_components/button/Button';
-import { Spinner } from '~/components/globals/Spinner';
+import { Spinner } from 'app/_components/globals/Spinner';
 
 import swell from '~/lib/SwellJS';
 import { notifyFailure, notifySuccess } from '~/utils/toastifies';
